Refactor: Platform-specific <-- method for Scala 2/3 compatibility
- Move SimpleKey to platform-specific dirs (scala-2.13/, scala-3/)
- Scala 3: Add [ThisV <: V] bound to prevent union type widening
- Scala 2: Keep [ThisV] unbounded (emulated unions can't prove subtyping)
- Use abstract bindSource method so subtypes stay in shared code
